+++ components/reporters/xml.rb
@@ -16,6 +16,7 @@ class Arachni::Reporters::XML < Arachni::Reporter::Base
     LOCAL_SCHEMA  = File.dirname( __FILE__ ) + '/xml/schema.xsd'
     REMOTE_SCHEMA = 'https://raw.githubusercontent.com/Arachni/arachni/' <<
         "v#{Arachni::VERSION}/components/reporters/xml/schema.xsd"
+    NULL          = '[ARACHNI_NULL]'
 
     def run
         builder = Nokogiri::XML::Builder.new do |xml|
@@ -31,7 +32,7 @@ class Arachni::Reporters::XML < Arachni::Reporter::Base
 
                 xml.sitemap {
                     report.sitemap.each do |url, code|
-                        xml.entry url: url, code: code
+                        xml.entry url: replace_nulls( url ), code: code
                     end
                 }
 
@@ -66,15 +67,15 @@ class Arachni::Reporters::XML < Arachni::Reporter::Base
                             xml.vector {
                                 xml.class_ vector.class
                                 xml.type vector.type
-                                xml.url vector.url
-                                xml.action vector.action
+                                xml.url replace_nulls( vector.url )
+                                xml.action replace_nulls( vector.action )
 
                                 if vector.respond_to? :source
-                                    xml.source vector.source
+                                    xml.source replace_nulls( vector.source )
                                 end
 
                                 if vector.respond_to? :seed
-                                    xml.seed vector.seed
+                                    xml.seed replace_nulls( vector.seed )
                                 end
 
                                 if issue.active?
@@ -82,7 +83,7 @@ class Arachni::Reporters::XML < Arachni::Reporter::Base
                                 end
 
                                 if vector.respond_to? :affected_input_name
-                                    xml.affected_input_name vector.affected_input_name
+                                    xml.affected_input_name replace_nulls( vector.affected_input_name )
                                 end
 
                                 if vector.respond_to? :inputs
@@ -169,6 +170,8 @@ class Arachni::Reporters::XML < Arachni::Reporter::Base
         end
 
         IO.binwrite( outfile, xml )
+
+        print_info "Null bytes have been replaced with: #{NULL}"
         print_status "Saved in '#{outfile}'."
     end
 
@@ -178,15 +181,19 @@ class Arachni::Reporters::XML < Arachni::Reporter::Base
             description:  %q{Exports the audit results as an XML (.xml) file.},
             content_type: 'text/xml',
             author:       'Tasos "Zapotek" Laskos <tasos.laskos@arachni-scanner.com>',
-            version:      '0.3.5',
+            version:      '0.3.6',
             options:      [ Options.outfile( '.xml' ), Options.skip_responses ]
         }
+    end
+
+    def replace_nulls( s )
+        s.to_s.gsub( "\0", NULL )
     end
 
     def add_inputs( xml, inputs, name = :inputs )
         xml.send( name ) {
             inputs.each do |k, v|
-                xml.input( name: k, value: v )
+                xml.input( name: replace_nulls( k ), value: replace_nulls( v ) )
             end
         }
     end
@@ -194,7 +201,7 @@ class Arachni::Reporters::XML < Arachni::Reporter::Base
     def add_headers( xml, headers )
         xml.headers {
             headers.each do |k, v|
-                xml.header( name: k, value: v )
+                xml.header( name: replace_nulls( k ), value: replace_nulls( v ) )
             end
         }
     end
@@ -202,30 +209,30 @@ class Arachni::Reporters::XML < Arachni::Reporter::Base
     def add_parameters( xml, parameters )
         xml.parameters {
             parameters.each do |k, v|
-                xml.parameter( name: k, value: v )
+                xml.parameter( name: replace_nulls( k ), value: replace_nulls( v ) )
             end
         }
     end
 
     def add_page( xml, page, name = :page )
         xml.send( name ) {
-            xml.body page.body
+            xml.body replace_nulls( page.body )
 
             request = page.request
             xml.request {
-                xml.url request.url
+                xml.url replace_nulls( request.url )
                 xml.method_ request.method
 
                 add_parameters( xml, request.parameters )
                 add_headers( xml, request.headers )
 
-                xml.body request.effective_body
-                xml.raw request.to_s
+                xml.body replace_nulls( request.effective_body )
+                xml.raw replace_nulls( request )
             }
 
             response = page.response
             xml.response {
-                xml.url response.url
+                xml.url replace_nulls( response.url )
                 xml.code response.code
                 xml.ip_address response.ip_address
                 xml.time response.time.round( 4 )
@@ -234,13 +241,13 @@ class Arachni::Reporters::XML < Arachni::Reporter::Base
 
                 add_headers( xml, response.headers )
 
-                xml.body response.body
-                xml.raw_headers response.headers_string
+                xml.body replace_nulls( response.body )
+                xml.raw_headers replace_nulls( response.headers_string )
             }
 
             dom = page.dom
             xml.dom {
-                xml.url dom.url
+                xml.url replace_nulls( dom.url )
 
                 xml.transitions {
                     dom.transitions.each do |transition|
@@ -257,8 +264,8 @@ class Arachni::Reporters::XML < Arachni::Reporter::Base
                         xml.data_flow_sink {
                             xml.object sink.object
                             xml.tainted_argument_index sink.tainted_argument_index
-                            xml.tainted_value sink.tainted_value
-                            xml.taint_ sink.taint
+                            xml.tainted_value replace_nulls( sink.tainted_value )
+                            xml.taint_ replace_nulls( sink.taint )
 
                             add_function( xml, sink.function )
                             add_trace( xml, sink.trace )